split `plot.py` into `{plot,tplot}.py`
